Create logLead module and update index
