+++ index.js
@@ -3,7 +3,7 @@ import { config } from 'dotenv';
 import { OpenAI } from 'openai';
 import twilio from 'twilio';
 import axios from 'axios';
-import { google } from 'googleapis';
+import { logLead } from './logLead.js';
 
 config();
 
@@ -14,13 +14,6 @@ app.use(express.json());
 const openai = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
 const twilioClient = twilio(process.env.TWILIO_ACCOUNT_SID, process.env.TWILIO_AUTH_TOKEN);
 const conversations = new Map();
-
-// Google Sheets setup
-const sheetsAuth = new google.auth.GoogleAuth({
-  keyFile: './service-account.json',
-  scopes: ['https://www.googleapis.com/auth/spreadsheets'],
-});
-const sheets = google.sheets({ version: 'v4', auth: sheetsAuth });
 
 const systemPrompt =
   "You're Daniel's AI assistant. A seller has just called in. Start the conversation by confirming who they are and asking if they’re open to a cash offer.";
@@ -132,34 +125,7 @@ app.get('/dealsync', async (req, res) => {
 });
 
 // === Lead Logging Helper ===
-export async function logLead(data) {
-  const {
-    phone = '',
-    address = '',
-    status = '',
-    summary = '',
-    tags = [],
-    callTime = new Date().toISOString(),
-    messages = [],
-  } = data || {};
-
-  const row = [
-    phone,
-    address,
-    status,
-    summary,
-    Array.isArray(tags) ? tags.join(',') : '',
-    new Date(callTime).toISOString(),
-    JSON.stringify(messages),
-  ];
-
-  await sheets.spreadsheets.values.append({
-    spreadsheetId: process.env.SHEET_ID,
-    range: 'Sheet1!A1',
-    valueInputOption: 'USER_ENTERED',
-    requestBody: { values: [row] },
-  });
-}
+// logLead imported from ./logLead.js
 
 // === Log Lead Route ===
 app.post('/log-lead', async (req, res) => {
